Make applinkactivity check if subreddit exists
