+++ app/src/main/res/layout/item_comment_loadmore.xml
@@ -4,7 +4,8 @@
     android:layout_height="wrap_content"
     android:clickable="true"
     android:foreground="@drawable/list_selector"
-    android:orientation="vertical">
+    android:orientation="vertical"
+    android:focusable="true">
 
     <TextView
         android:id="@+id/comment_loadmore"
