+++ app/src/main/res/layout/item_post.xml
@@ -50,7 +50,8 @@
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintRight_toRightOf="parent"
             app:layout_constraintTop_toTopOf="@+id/post_messages"
-            android:layout_marginTop="8dp" />
+            android:layout_marginTop="8dp"
+            android:focusable="true" />
 
         <Button
             android:id="@+id/post_messages"
@@ -71,7 +72,8 @@
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintRight_toLeftOf="@+id/post_open_browser"
             android:layout_marginTop="8dp"
-            app:layout_constraintTop_toBottomOf="@+id/post_domain" />
+            app:layout_constraintTop_toBottomOf="@+id/post_domain"
+            android:focusable="true" />
 
         <Button
             android:id="@+id/post_likes"
