+++ app/src/main/res/layout/item_post_small.xml
@@ -50,7 +50,8 @@
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintDimensionRatio=""
             app:layout_constraintLeft_toRightOf="@+id/post_messages"
-            android:layout_marginLeft="16dp" />
+            android:layout_marginLeft="16dp"
+            android:focusable="true" />
 
         <Button
             android:id="@+id/post_messages"
@@ -71,7 +72,8 @@
             app:layout_constraintTop_toBottomOf="@+id/post_domain"
             app:layout_constraintVertical_bias="0.0"
             android:layout_marginLeft="16dp"
-            app:layout_constraintLeft_toLeftOf="parent" />
+            app:layout_constraintLeft_toLeftOf="parent"
+            android:focusable="true" />
 
         <TextView
             android:id="@+id/post_title"
